+++ pages/index.jsx
@@ -1,69 +1,75 @@
-import Head from "next/head";
-import Image from "next/image";
-import styles from "../styles/Home.module.css";
+import { useState } from "react";
+import styles from "./login.module.scss";
+import axios from "axios";
+import router from "next/router";
+import { BlueBtn } from "../components/login/btn/mainBtn";
+import InputBox from "../components/login/btn/inputBox";
+import TextBtn from "../components/login/btn/textBtn";
+import { ImageLogo, NameLogo } from "../components/login/logos";
 
-export default function Home() {
+const Login = ({}) => {
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+
+  const checkAccount = () => {
+    if (username === "" || password === "") {
+      return alert("id와 password를 입력해주세요");
+    }
+
+    const LoginRequest = new Object();
+    LoginRequest.username = username;
+    LoginRequest.password = password;
+
+    axios
+      .post("/login", LoginRequest)
+      .then(function (response) {
+        console.log(response);
+        const accessToken = response.data.split(" ")[1];
+        localStorage.setItem("accessToken", accessToken);
+        alert("로그인에 성공하였습니다");
+        router.push("/myclass");
+      })
+      .catch(function (error) {
+        alert("로그인에 실패했습니다.");
+        console.log(error);
+      });
+  };
+  const onChangeUsername = (e) => {
+    setUsername(e.target.value);
+  };
+  const onChangePassword = (e) => {
+    setPassword(e.target.value);
+  };
+
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{" "}
-          <code className={styles.code}>pages/index.js</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/master/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{" "}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
+    <section className={styles.loginSection}>
+      <h1 className={styles.title}>{"튜터로 로그인"}</h1>
+      <span className={styles.imageLogo}>
+        <ImageLogo />
+      </span>
+      <div className={styles.btns}>
+        <InputBox
+          type={"email"}
+          placeholder={"ID(Email)"}
+          onChange={onChangeUsername}
+          value={username}
+        />
+        <InputBox
+          type={"password"}
+          placeholder={"Password"}
+          onChange={onChangePassword}
+          value={password}
+        />
+        <BlueBtn text={"로그인 하기"} type="submit" onClick={checkAccount} />
+        <span className={styles.textBtn}>
+          <TextBtn text={"아이디 찾기"} />
+          <span className={styles.text}>|</span>
+          <TextBtn text={"비밀번호 찾기"} />
+        </span>
+      </div>
+      <NameLogo />
+    </section>
   );
-}
+};
+
+export default Login;
